user.spec.rb追加
